checked BIGINT to INT

+++ alx_book_store.sql
@@ -26,16 +26,16 @@ CREATE TABLE IF NOT EXISTS Customers (
 );
 
 CREATE TABLE IF NOT EXISTS Orders (
-    order_id BIGINT PRIMARY KEY AUTO_INCREMENT,
-    customer_id BIGINT,
+    order_id INT PRIMARY KEY AUTO_INCREMENT,
+    customer_id INT,
     order_date DATE,
     FOREIGN KEY (customer_id) REFERENCES Customers(customer_id)
 );
 
 CREATE TABLE IF NOT EXISTS Order_Details (
     orderdetailid BIGINT PRIMARY KEY AUTO_INCREMENT,
-    order_id BIGINT,
-    book_id BIGINT,
+    order_id INT,
+    book_id INT,
     quantity DOUBLE,
     FOREIGN KEY (order_id) REFERENCES Orders(order_id),
     FOREIGN KEY (book_id) REFERENCES Books(book_id)
